예매페이지 연결중

+++ back/hrw/src/main/java/com/cinema/hrw/entity/OrderEntity.java
@@ -62,7 +62,7 @@ public class OrderEntity {
 
 
     @Transient
-    private int allPrice; // 총결제액
+    private Long allPrice; // 총결제액
 
 
 }
